Create Firewall
Allow incoming traffic on port 53 (DNS) to allow DNS resolution
Allow incoming traffic from the local network (10.0.0.0/8)
Log dropped packets with a prefix "Firewall: "
In this script, allowing traffic on port 53 for DNS resolution, since without that the server will not be able to resolve the domain names and will not be able to access the internet. We also added an exception for traffic coming from the local network( 10.0.0.0/8) this will allow communication between the machines in the same network. Lastly, a rule to log the dropped packets with a prefix to know what kind of traffic is being blocked by the firewall.